scroll content widget size policy switcher. storage data type to int64.

+++ qofia-ffi/room_list_item.go
@@ -45,13 +45,13 @@ func NewMessageForGroup(jso *simplejson.Json) *Message {
 	peerName := jso.Get("margs").GetIndex(0).MustString()
 	groupTitle := jso.Get("margs").GetIndex(2).MustString()
 	_ = groupTitle
-	eventId := gopp.MustInt(jso.Get("margs").GetIndex(4).MustString())
+	eventId := int64(gopp.MustInt(jso.Get("margs").GetIndex(4).MustString()))
 
 	this := &Message{}
 	this.Msg = message
 	this.Peer = peerName
 	this.Time = time.Now()
-	this.EventId = int64(eventId)
+	this.EventId = eventId
 
 	this.refmtmsg()
 	return this
@@ -62,13 +62,13 @@ func NewMessageForFriend(jso *simplejson.Json) *Message {
 	fname := jso.Get("margs").GetIndex(0).MustString()
 	pubkey := jso.Get("margs").GetIndex(1).MustString()
 	_, _, _ = msg, fname, pubkey
-	eventId := gopp.MustInt(jso.Get("margs").GetIndex(2).MustString())
+	eventId := int64(gopp.MustInt(jso.Get("margs").GetIndex(2).MustString()))
 
 	this := &Message{}
 	this.Msg = msg
 	this.Peer = fname
 	this.Time = time.Now()
-	this.EventId = int64(eventId)
+	this.EventId = eventId
 
 	this.refmtmsg()
 	return this
